+++ packages/conjure-client/src/fetchBridge/__tests__/fetchBridgeTests.ts
@@ -18,7 +18,8 @@
 import * as fetchMock from "fetch-mock";
 import { ConjureError, ConjureErrorType } from "../../errors";
 import { IHttpApiBridge, IHttpEndpointOptions, MediaType } from "../../httpApiBridge";
-import { FetchBridge, IUserAgent } from "../fetchBridge";
+import { IUserAgent } from "../../userAgent";
+import { FetchBridge } from "../fetchBridge";
 
 const baseUrl = "https://host.domain/path";
 const token = "TOKEN";
@@ -482,7 +483,7 @@ function createFetchRequest(opts: ICreateFetchRequestOpts): RequestInit {
         headers: {
             ...headers,
             Authorization: `Bearer ${token}`,
-            "Fetch-User-Agent": "foo/1.2.3",
+            "Fetch-User-Agent": "foo/1.2.3 conjure-client/0.0.0",
         },
         method,
     };
